server/requirements.txt
Flask
gunicorn
Pillow==5.3.0
aiohttp
requests
boto3